Cleanup event listeners components
